requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
twilio==8.13.0
openai==1.12.0
python-dotenv==1.0.1
jinja2==3.1.3
aiosqlite==0.19.0
python-multipart==0.0.9
httpx==0.26.0
